+++ static_integration_test.go
@@ -7,7 +7,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"tidbyt.dev/gtfs"
 	"tidbyt.dev/gtfs/model"
 	"tidbyt.dev/gtfs/testutil"
 )
@@ -116,7 +115,7 @@ func testGTFSStaticIntegrationDepartures(t *testing.T, backend string) {
 
 	// Feb 3rd is a Monday
 	departures, _ := g.Departures("G33S", time.Date(2020, 2, 3, 22, 50, 0, 0, tz), 10*time.Minute, -1, "", -1, nil)
-	assert.Equal(t, []gtfs.Departure{
+	assert.Equal(t, []model.Departure{
 		{
 			StopID:       "G33S",
 			RouteID:      "G",
@@ -139,7 +138,7 @@ func testGTFSStaticIntegrationDepartures(t *testing.T, backend string) {
 
 	// Feb 17 is also a Monday, but President's Day
 	departures, _ = g.Departures("G33S", time.Date(2020, 2, 17, 22, 50, 0, 0, tz), 10*time.Minute, -1, "", -1, nil)
-	assert.Equal(t, []gtfs.Departure{
+	assert.Equal(t, []model.Departure{
 		{
 			StopID:       "G33S",
 			RouteID:      "G",
@@ -155,7 +154,7 @@ func testGTFSStaticIntegrationDepartures(t *testing.T, backend string) {
 	// reverse order in stop_times.txt, but will be still be
 	// returned ordered by departure time.
 	departures, _ = g.Departures("G33S", time.Date(2020, 2, 17, 22, 50, 0, 0, tz), 13*time.Minute, -1, "", -1, nil)
-	assert.Equal(t, []gtfs.Departure{
+	assert.Equal(t, []model.Departure{
 		{
 			StopID:       "G33S",
 			RouteID:      "G",
@@ -178,7 +177,7 @@ func testGTFSStaticIntegrationDepartures(t *testing.T, backend string) {
 
 	// Feb 16 is a Sunday
 	departures, _ = g.Departures("G33S", time.Date(2020, 2, 16, 22, 50, 0, 0, tz), 10*time.Minute, -1, "", -1, nil)
-	assert.Equal(t, []gtfs.Departure{
+	assert.Equal(t, []model.Departure{
 		{
 			StopID:       "G33S",
 			RouteID:      "G",
